Update fun_Transit.cpp
Equation NumWorkers
Fixed the computation of vacancies, formerly computed as the ratio of vacancies divided by the number of new recruits. Replaced with the number of workers.

Equation TotKLaborForce
Replaced SUM(“LaborForce”) with SUM(“KLaborForce”)

Equation TotKNbrWorkers
Removed. Wrong, and never used anyway

Comparison NbrWorkers vs TotLaborForce+TotKLaborForce
The two values should be identical, but they are not. To assess the motivation I advanced the computation of Demography to be done before NbrWorkers. Got a deadlock because of MinWage, so removed the call (without using the value) to the computation of NbrWorkers. The result now match but for a (small) irregular approximation of exactly 10, 20 or 0 units. The error persists also when placing Demography to be computed before TotXLaborForce’s.